+++ src/tokens.ts
@@ -4,10 +4,10 @@ import {InjectionToken} from '@angular/core';
  * An injection token representing the local storage.
  * @type {InjectionToken}
  */
-export const LocalStorage = new InjectionToken('LocalStorage');
+export const localStorage = new InjectionToken('storage.local');
 
 /**
  * An injection token representing the session storage.
  * @type {InjectionToken}
  */
-export const SessionStorage = new InjectionToken('SessionStorage');
+export const sessionStorage = new InjectionToken('storage.session');